Move get_or_create_spark_session into execution engine
